+++ src/appwindow.cpp
@@ -162,7 +162,7 @@ AppWindow::AppWindow(bool cpp, bool java, bool python, bool noRestoreSession, in
 
 void AppWindow::finishConstruction()
 {
-    if (ui->tabWidget->count() == 0)
+    if (tabCount() == 0)
         openTab("");
 
 #ifdef Q_OS_WIN
@@ -199,7 +199,7 @@ AppWindow::~AppWindow()
 {
     LOG_INFO("Destruction started");
     saveSettings();
-    while (ui->tabWidget->count())
+    while (tabCount())
     {
         auto *tmp = ui->tabWidget->widget(0);
         ui->tabWidget->removeTab(0);
@@ -269,7 +269,7 @@ void AppWindow::changeEvent(QEvent *event)
 /******************** PRIVATE METHODS ********************/
 void AppWindow::setConnections()
 {
-    connect(ui->tabWidget, &QTabWidget::tabCloseRequested, this, &AppWindow::onTabCloseRequested);
+    connect(ui->tabWidget, &QTabWidget::tabCloseRequested, this, [this](int index) { closeTab(index); });
     connect(ui->tabWidget, &QTabWidget::currentChanged, this, &AppWindow::onTabChanged);
     ui->tabWidget->tabBar()->setContextMenuPolicy(Qt::CustomContextMenu);
     connect(ui->tabWidget->tabBar(), &QTabBar::customContextMenuRequested, this, &AppWindow::onTabContextMenuRequested);
@@ -397,11 +397,16 @@ void AppWindow::maybeSetHotkeys()
     }
 }
 
-bool AppWindow::closeTab(int index)
+bool AppWindow::closeTab(int index, bool noConfirmQuit)
 {
     LOG_INFO(INFO_OF(index));
     auto *tmp = windowAt(index);
-    if (tmp->closeConfirm())
+    if (!tmp)
+    {
+        LOG_WARN("Current window is nullptr");
+        return false;
+    }
+    if (noConfirmQuit || tmp->closeConfirm())
     {
         ui->tabWidget->removeTab(index);
         onEditorFileChanged();
@@ -409,6 +414,11 @@ bool AppWindow::closeTab(int index)
         return true;
     }
     return false;
+}
+
+bool AppWindow::closeWindow(MainWindow *window, bool noConfirmQuit)
+{
+    return closeTab(indexOfWindow(window), noConfirmQuit);
 }
 
 void AppWindow::saveSettings()
@@ -431,9 +441,8 @@ void AppWindow::openTab(MainWindow *window, MainWindow *after)
     connect(window, &MainWindow::compileOrRunTriggered, this, &AppWindow::onCompileOrRunTriggered);
     connect(window, &MainWindow::fileSaved, this, &AppWindow::onFileSaved);
 
-    ui->tabWidget->setCurrentIndex(
-        ui->tabWidget->insertTab(after ? ui->tabWidget->indexOf(after) + 1 : ui->tabWidget->currentIndex() + 1, window,
-                                 window->getTabTitle(false, true)));
+    setTabAt(ui->tabWidget->insertTab(after ? indexOfWindow(after) + 1 : ui->tabWidget->currentIndex() + 1, window,
+                                      window->getTabTitle(false, true)));
 
     window->getEditor()->setFocus();
     onEditorFileChanged();
@@ -550,7 +559,7 @@ bool AppWindow::quit()
     {
         LOG_INFO("quit() called without hotExit");
         on_actionCloseAll_triggered();
-        if (ui->tabWidget->count() >= 1)
+        if (tabCount() >= 1)
         {
             LOG_INFO("Closing is cancelled");
             return false;
@@ -568,7 +577,7 @@ int AppWindow::getNewUntitledIndex()
 {
     int index = 0;
     QSet<int> vis;
-    for (int t = 0; t < ui->tabWidget->count(); ++t)
+    for (int t = 0; t < tabCount(); ++t)
     {
         auto *tmp = windowAt(t);
         if (tmp->isUntitled() && tmp->getProblemURL().isEmpty())
@@ -713,7 +722,7 @@ void AppWindow::on_actionSaveAs_triggered()
 
 void AppWindow::on_actionSaveAll_triggered()
 {
-    for (int t = 0; t < ui->tabWidget->count(); ++t)
+    for (int t = 0; t < tabCount(); ++t)
     {
         auto *tmp = windowAt(t);
         if (!tmp->save(true, tr("Save All")))
@@ -730,7 +739,7 @@ void AppWindow::on_actionCloseCurrent_triggered()
 
 void AppWindow::on_actionCloseAll_triggered()
 {
-    for (int t = 0; t < ui->tabWidget->count(); t++)
+    for (int t = 0; t < tabCount(); t++)
     {
         if (closeTab(t))
             --t;
@@ -741,7 +750,7 @@ void AppWindow::on_actionCloseAll_triggered()
 
 void AppWindow::on_actionCloseSaved_triggered()
 {
-    for (int t = 0; t < ui->tabWidget->count(); t++)
+    for (int t = 0; t < tabCount(); t++)
         if (!windowAt(t)->isTextChanged() && closeTab(t))
             --t;
 }
@@ -869,11 +878,6 @@ bool AppWindow::forceClose()
     return close();
 }
 
-void AppWindow::onTabCloseRequested(int index)
-{
-    closeTab(index);
-}
-
 void AppWindow::onTabChanged(int index)
 {
     LOG_INFO(INFO_OF(index));
@@ -923,6 +927,7 @@ void AppWindow::onTabChanged(int index)
         connect(tmp->getRightSplitter(), &QSplitter::splitterMoved, this, &AppWindow::onRightSplitterMoved);
 
     triggerWakaTime(tmp);
+    tmp->getEditor()->setFocus();
 }
 
 void AppWindow::onEditorFileChanged()
@@ -931,7 +936,7 @@ void AppWindow::onEditorFileChanged()
     {
         QMap<QString, QVector<int>> tabsByName;
 
-        for (int t = 0; t < ui->tabWidget->count(); ++t)
+        for (int t = 0; t < tabCount(); ++t)
         {
             tabsByName[windowAt(t)->getTabTitle(false, false)].push_back(t);
         }
@@ -969,7 +974,7 @@ void AppWindow::onEditorFileChanged()
 
 void AppWindow::onEditorTextChanged(MainWindow *window)
 {
-    int index = ui->tabWidget->indexOf(window);
+    int index = indexOfWindow(window);
     if (index != -1)
     {
         auto title = ui->tabWidget->tabText(index);
@@ -1054,7 +1059,7 @@ void AppWindow::onSettingsApplied(const QString &pagePath)
 {
     LOG_INFO("Apply settings for " << INFO_OF(pagePath));
 
-    for (int i = 0; i < ui->tabWidget->count(); ++i)
+    for (int i = 0; i < tabCount(); ++i)
     {
         windowAt(i)->applySettings(pagePath);
         onEditorTextChanged(windowAt(i));
@@ -1127,11 +1132,11 @@ void AppWindow::onIncomingCompanionRequest(const Extensions::CompanionData &data
 {
     LOG_INFO("Request from competitive companion arrived");
 
-    for (int i = 0; i < ui->tabWidget->count(); ++i)
+    for (int i = 0; i < tabCount(); ++i)
     {
         if (windowAt(i)->getProblemURL() == data.url)
         {
-            ui->tabWidget->setCurrentIndex(i);
+            setTabAt(i);
             currentWindow()->applyCompanion(data);
             return;
         }
@@ -1187,18 +1192,18 @@ void AppWindow::onRightSplitterMoved()
     SettingsHelper::setRightSplitterSize(splitter->saveState());
 }
 
-void AppWindow::openTab(const QString &path, MainWindow *after)
+void AppWindow::openTab(const QString &path, QString lang, MainWindow *after)
 {
     LOG_INFO("OpenTab Path is " << path);
     if (!path.isEmpty())
     {
         auto fileInfo = QFileInfo(path);
-        for (int t = 0; t < ui->tabWidget->count(); t++)
+        for (int t = 0; t < tabCount(); t++)
         {
             auto tPath = qobject_cast<MainWindow *>(ui->tabWidget->widget(t))->getFilePath();
             if (path == tPath || (fileInfo.exists() && fileInfo == QFileInfo(tPath)))
             {
-                ui->tabWidget->setCurrentIndex(t);
+                setTabAt(t);
                 return;
             }
         }
@@ -1206,16 +1211,18 @@ void AppWindow::openTab(const QString &path, MainWindow *after)
 
     auto *newWindow = new MainWindow(path, getNewUntitledIndex(), this);
 
-    QString lang = SettingsHelper::getDefaultLanguage();
+    if (lang.isEmpty())
+    {
 
-    auto suffix = QFileInfo(path).suffix();
+        auto suffix = QFileInfo(path).suffix();
 
-    if (Util::cppSuffix.contains(suffix))
-        lang = "C++";
-    else if (Util::javaSuffix.contains(suffix))
-        lang = "Java";
-    else if (Util::pythonSuffix.contains(suffix))
-        lang = "Python";
+        if (Util::cppSuffix.contains(suffix))
+            lang = "C++";
+        else if (Util::javaSuffix.contains(suffix))
+            lang = "Java";
+        else if (Util::pythonSuffix.contains(suffix))
+            lang = "Python";
+    }
 
     newWindow->setLanguage(lang);
 
@@ -1447,9 +1454,9 @@ void AppWindow::on_actionToggleBlockComment_triggered()
 
 void AppWindow::onConfirmTriggered(MainWindow *widget)
 {
-    int index = ui->tabWidget->indexOf(widget);
+    int index = indexOfWindow(widget);
     if (index != -1)
-        ui->tabWidget->setCurrentIndex(index);
+        setTabAt(index);
 }
 
 void AppWindow::onTabContextMenuRequested(const QPoint &pos)
@@ -1467,19 +1474,19 @@ void AppWindow::onTabContextMenuRequested(const QPoint &pos)
         tabMenu->addAction(tr("Close"), [index, this] { closeTab(index); });
 
         tabMenu->addAction(tr("Close Others"), [window, this] {
-            for (int i = 0; i < ui->tabWidget->count(); ++i)
+            for (int i = 0; i < tabCount(); ++i)
                 if (windowAt(i) != window && closeTab(i))
                     --i;
         });
 
         tabMenu->addAction(tr("Close to the Left"), [window, this] {
-            for (int i = 0; i < ui->tabWidget->count() && windowAt(i) != window; ++i)
+            for (int i = 0; i < tabCount() && windowAt(i) != window; ++i)
                 if (closeTab(i))
                     --i;
         });
 
         tabMenu->addAction(tr("Close to the Right"), [index, this] {
-            for (int i = index + 1; i < ui->tabWidget->count(); ++i)
+            for (int i = index + 1; i < tabCount(); ++i)
                 if (closeTab(i))
                     --i;
         });
@@ -1560,6 +1567,20 @@ void AppWindow::onTabContextMenuRequested(const QPoint &pos)
     }
 }
 
+void AppWindow::setTabAt(int index)
+{
+    ui->tabWidget->setCurrentIndex(index);
+}
+
+int AppWindow::indexOfWindow(MainWindow *window)
+{
+    return ui->tabWidget->indexOf(window);
+}
+
+int AppWindow::tabCount() const
+{
+    return ui->tabWidget->count();
+}
 MainWindow *AppWindow::currentWindow()
 {
     int current = ui->tabWidget->currentIndex();